+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 v5.0.0
 ----
 * Support Capacitor Projects (Beta)
+* Support Yarn Projects (Beta)
 
 #### Breaking changes
 * Modified `isMonacaProject` function. The project is "Monaca" if it is either "Cordova" or "Capacitor" project.